+++ src/index.js
@@ -1,22 +1,47 @@
 import React from 'react';
 import Editor from 'draft-js-plugins-editor';
-import {EditorState} from 'draft-js';
+import Draft, {
+  ContentState,
+  EditorState,
+  convertToRaw
+} from 'draft-js';
 import createMarkdownShortcutsPlugin from 'draft-js-markdown-shortcuts-plugin';
 import Immutable from 'immutable';
-import './index.less';
+import 'prismjs/themes/prism-funky.css';
+import prismPlugin from './plugins/prism';
+// import codePlugin from './plugins/code';
+import customStyleMap from './styles';
+// import './styles/prism.less';
 
+window.Draft = Draft;
 const plugins = [
+  // prismPlugin,
   createMarkdownShortcutsPlugin()
 ];
+const contentState = ContentState.createFromText('');
+const initialEditorState = EditorState.createWithContent(contentState);
+const rawContentState = convertToRaw(initialEditorState.getCurrentContent());
+
 class MabyEditor extends React.Component {
   constructor(props) {
     super(props);
-    this.state = { editorState: EditorState.createEmpty() };
-    this.focus = () => this.refs.editor.focus();
+    this.state = { editorState: initialEditorState, rawContent: rawContentState};
+    this.Draft = Draft;
+  }
+  componentDidMount = () => {
+    const { editor } = this;
+    if (editor) {
+      setTimeout(editor.focus.bind(editor), 1000);
+    }
   }
   onChange = (editorState) => {
-    this.setState({ editorState })
-    console.log('editorState', editorState);
+    window.editorState = editorState;
+    window.rawContent = convertToRaw(editorState.getCurrentContent());
+    this.setState({ 
+      editorState,
+      rawContent: convertToRaw(editorState.getCurrentContent())
+    })
+    // console.log('editorState', convertToRaw(editorState.getCurrentContent()));
   };
   render() {
     const { editorState } = this.state;
@@ -26,8 +51,10 @@ class MabyEditor extends React.Component {
         <Editor
           editorState={editorState}
           onChange={this.onChange}
+          customStyleMap={customStyleMap}
           plugins={plugins}
-          ref="editor"
+          spellCheck
+          ref={(element) => { this.editor = element; }}
         />
       </div>
     );