+++ src/styles/block.js
@@ -1,0 +1,10 @@
+const fontSize = '16px';
+export default {
+  'CODE': {
+    fontFamily: 'monospace',
+    fontSize: fontSize,
+    borderLeft: '5px solid #eee',
+    color: 'red',
+    padding: '8px'
+  }
+};