+++ src/styles/index.js
@@ -1,0 +1,3 @@
+import inline from './inline';
+import block from './block';
+export default { ...inline, ...block};