+++ src/styles/inline.js
@@ -1,0 +1,5 @@
+export default {
+  'BOLD': {
+    color: 'red'
+  }
+};